moved generate diff input fixture to conftest, for usage in other tests
